backend/requirements.txt
Django>=4.2,<6.1
djangorestframework>=3.14
django-cors-headers>=4.0

gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
psycopg2-binary>=2.9

sentry-sdk>=2.0
